ducterry - 28.02.2022 - Update Trace Log

+++ src/main/java/com/ducterry/base/entity/log/AppTraceLog.java
@@ -17,7 +17,7 @@ public class AppTraceLog {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "ID")
-    private BigDecimal id;
+    private Long  id;
     @Column(name = "CREATED_DATE")
     private Date createdDate;
     @Column(name = "BODY_REQUEST")
